WIP: creating TIFReader class
