+++ ci/__main__.py
@@ -28,8 +28,29 @@ python_version_option = click.option(
 	help="Python version for environment"
 )
 
+@cli.command(name="build-env", help="Creates the execution environment")
+@python_version_option
+def build_env(python_version):
+    env_name = get_env_name(python_version)
+    check_call([
+        "edm", "envs", "remove", "--purge", "--force",
+        "--yes", env_name])
+    check_call(
+        ["edm", "envs", "create", "--version", python_version,
+         env_name])
+
+    check_call([
+        "edm", "install", "-e", env_name,
+        "--yes"] + CORE_DEPS + DEV_DEPS + DOCS_DEPS)
+
+    if len(PIP_DEPS):
+        check_call([
+            "edm", "run", "-e", env_name, "--",
+            "pip", "install"] + PIP_DEPS)
+
+
 @cli.command(name="install",
-			 help='Creates the execution environment and binaries')
+			 help='Creates the execution binary inside the PyFibre environment')
 @python_version_option
 def install(python_version):
 
@@ -52,14 +73,23 @@ def uninstall(python_version):
 	edm_run(env_name, ['pip', 'uninstall', '-e', '.'])
 
 
-@cli.command(name="test",
-			 help='Runs unit tests for repository')
+@cli.command(help="Run flake")
+@python_version_option
+def flake8(python_version):
+	env_name = get_env_name()
+
+	check_call(["edm", "run", "-e", env_name, "--", "flake8", "."])
+
+
+@cli.command(help="Run the tests")
 @python_version_option
 def test(python_version):
+    env_name = get_env_name()
 
-	env_name = get_env_name()
-	edm_run(env_name, ['python', '-m', 'unittest',
-					   'discover', '*/test/*'])
+    check_call([
+        "edm", "run", "-e", env_name, "--", "python", "-m", "unittest",
+        "discover"
+    ])
 
 
 if __name__ == "__main__":
