+++ pyfibre/tests/stubs/test-pyfibre-pl-shg-Stack.tif
