+++ pyfibre/tools/filters.py
