+++ pyfibre/utilities.py
@@ -11,9 +11,8 @@ Last Modified: 12/04/2018
 import logging
 import numpy as np
 
-import sys, os, pickle
+import pickle
 
-from skimage import img_as_float, io
 
 SQRT3 = np.sqrt(3)
 SQRT2 = np.sqrt(2)
@@ -83,14 +82,6 @@ def check_file_name(file_name, file_type="", extension=""):
 	file_name = check_string(file_name, -1, '_', file_type)
 	
 	return file_name
-
-
-def load_image(image_name):
-
-	print(f"Loading {image_name}")
-	image = io.imread(image_name).astype(np.float64)
-
-	return image
 
 
 def save_npy(file_path, array):
